+++ editor/src/components/canvas/canvas-strategies/absolute-resize-delta-strategy.tsx
@@ -1,30 +1,19 @@
-import { isHorizontalPoint } from 'utopia-api/core'
-import { getLayoutProperty } from '../../../core/layout/getLayoutProperty'
-import { framePointForPinnedProp } from '../../../core/layout/layout-helpers-new'
 import { MetadataUtils } from '../../../core/model/element-metadata-utils'
-import { mapDropNulls } from '../../../core/shared/array-utils'
-import { isRight, right } from '../../../core/shared/either'
 import { ElementInstanceMetadataMap, JSXElement } from '../../../core/shared/element-template'
-import { CanvasRectangle, CanvasVector, offsetPoint } from '../../../core/shared/math-utils'
+import { CanvasVector, offsetPoint } from '../../../core/shared/math-utils'
 import { ElementPath } from '../../../core/shared/project-file-types'
 import { withUnderlyingTarget } from '../../editor/store/editor-state'
-import { stylePropPathMappingFn } from '../../inspector/common/property-path-hooks'
 import { EdgePosition } from '../canvas-types'
-import {
-  AdjustCssLengthProperty,
-  adjustCssLengthProperty,
-} from '../commands/adjust-css-length-command'
 import { setSnappingGuidelines } from '../commands/set-snapping-guidelines-command'
 import { updateHighlightedViews } from '../commands/update-highlighted-views-command'
 import { AbsoluteResizeControl } from '../controls/select-mode/absolute-resize-control'
-import { AbsolutePin } from './absolute-resize-helpers'
 import { GuidelineWithSnappingVector } from '../guideline'
 import { CanvasStrategy } from './canvas-strategy-types'
+import { getMultiselectBounds } from './shared-absolute-move-strategy-helpers'
 import {
-  getMultiselectBounds,
   resizeBoundingBox,
   runLegacyAbsoluteResizeSnapping,
-} from './shared-absolute-move-strategy-helpers'
+} from './shared-absolute-resize-strategy-helpers'
 import { createResizeCommands } from './shared-absolute-resize-strategy-helpers'
 
 export const absoluteResizeDeltaStrategy: CanvasStrategy = {
@@ -138,7 +127,7 @@ function snapDrag(
     drag,
     edgePosition,
     null,
-    false,
+    'non-center-based',
   )
   const { snapDelta, guidelinesWithSnappingVector } = runLegacyAbsoluteResizeSnapping(
     selectedElements,
@@ -147,7 +136,7 @@ function snapDrag(
     resizedUnsnappedBounds,
     canvasScale,
     lockedAspectRatio,
-    false,
+    'non-center-based',
   )
   const snappedDragVector = offsetPoint(drag, snapDelta)
